backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.21
sqlalchemy==2.0.34
psycopg[binary]==3.2.13
httpx==0.27.2
PyJWT>=2.10.1
openai==1.57.0
sentence-transformers>=2.2.2
cachetools>=5.3.0  # For JWT token caching
numpy>=1.24.0  # For cosine similarity calculations
requests>=2.31.0  # For web scraping and search APIs
beautifulsoup4>=4.12.0  # For HTML parsing
lxml>=4.9.0  # Better HTML parser for BeautifulSoup
PyPDF2>=3.0.0  # For PDF text extraction
python-docx>=1.1.0  # For Word document processing
openpyxl>=3.1.0  # For Excel file processing
pandas>=2.0.0  # For CSV and Excel data processing
Pillow>=10.0.0  # For image processing
tweepy>=4.14.0  # For Twitter API integration
newsapi-python>=0.2.7  # For News API integration
supabase>=2.27.0  # For Supabase storage integration

